fix(omnivoice): upcast fc2 output to float32 before acoustic decoder
fc2 weights are stored as float16 in the safetensors checkpoint, so the
.to(fc2.weight.dtype) cast makes fc2 run in fp16. The downstream DAC
acoustic decoder amplifies values beyond the fp16 range (~65504) via
ConvTranspose1d upsampling, producing NaN waveforms.

Fix by upcasting fc2 output back to float32 (.float()) before the
acoustic decoder so the upsampling always runs in full precision.

+++ vllm_omni/model_executor/models/omnivoice/omnivoice_decoder.py
@@ -112,8 +112,10 @@ class OmniVoiceDecoder(nn.Module):
                     quantized.isnan().any().item())
 
         # Project: [B, 1024, T] → fc2 → [B, 256, T]
-        # Cast to fc2 weight dtype to handle float32 RVQ output under fp16/bf16 engine
-        quantized = self.fc2(quantized.transpose(1, 2).to(self.fc2.weight.dtype)).transpose(1, 2)
+        # Cast to fc2 weight dtype (may be fp16 when weights stored as fp16),
+        # then upcast back to float32 — acoustic decoder uses ConvTranspose1d
+        # upsampling whose intermediate values exceed fp16 range (~65504), causing NaN.
+        quantized = self.fc2(quantized.transpose(1, 2).to(self.fc2.weight.dtype)).transpose(1, 2).float()
         logger.info("[DEBUG-DEC] after fc2: shape=%s dtype=%s std=%.4f max_abs=%.4f nan=%s",
                     tuple(quantized.shape), quantized.dtype,
                     quantized.std().item(), quantized.abs().max().item(),
